Rename directory opm/core/tof -> opm/core/flowdiagnostics.
