+++ opm/core/flowdiagnostics/AnisotropicEikonal.cpp
@@ -17,7 +17,7 @@
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include <opm/core/tof/AnisotropicEikonal.hpp>
+#include <opm/core/flowdiagnostics/AnisotropicEikonal.hpp>
 #include <opm/core/grid/GridUtilities.hpp>
 #include <opm/core/grid.h>
 #include <opm/core/utility/RootFinders.hpp>
