+++ opm/core/flowdiagnostics/AnisotropicEikonal.hpp
