+++ opm/core/flowdiagnostics/TofDiscGalReorder.cpp
@@ -20,8 +20,8 @@
 #include "config.h"
 #include <opm/core/grid/CellQuadrature.hpp>
 #include <opm/core/grid/FaceQuadrature.hpp>
-#include <opm/core/tof/TofDiscGalReorder.hpp>
-#include <opm/core/tof/DGBasis.hpp>
+#include <opm/core/flowdiagnostics/TofDiscGalReorder.hpp>
+#include <opm/core/flowdiagnostics/DGBasis.hpp>
 #include <opm/core/grid.h>
 #include <opm/core/utility/ErrorMacros.hpp>
 #include <opm/core/utility/SparseTable.hpp>
