+++ opm/core/flowdiagnostics/TofReorder.cpp
@@ -18,7 +18,7 @@
 */
 
 #include "config.h"
-#include <opm/core/tof/TofReorder.hpp>
+#include <opm/core/flowdiagnostics/TofReorder.hpp>
 #include <opm/core/grid.h>
 #include <opm/core/utility/ErrorMacros.hpp>
 #include <opm/core/utility/SparseTable.hpp>
